frontend/pyproject.toml
[project]
name = "messenger-ui"
version = "0.1.0"
description = "A user interface for my messenger application"
authors = [
    {name = "zinuhci",email = "nyekachigreat46@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "anyio (>=4.11.0,<5.0.0)",
    "backports-tarfile (>=1.2.0,<2.0.0)",
    "bidict (>=0.23.1,<0.24.0)",
    "build (>=1.3.0,<2.0.0)",
    "cachecontrol (>=0.14.3,<0.15.0)",
    "certifi (>=2025.10.5,<2026.0.0)",
    "charset-normalizer (>=3.4.4,<4.0.0)",
    "colorama (>=0.4.6,<0.5.0)",
    "distlib (>=0.4.0,<0.5.0)",
    "dulwich (>=0.24.7,<0.25.0)",
    "exceptiongroup (>=1.3.0,<2.0.0)",
    "fastjsonschema (>=2.21.2,<3.0.0)",
    "filelock (>=3.20.0,<4.0.0)",
    "findpython (>=0.7.0,<0.8.0)",
    "h11 (>=0.16.0,<0.17.0)",
    "httpcore (>=1.0.9,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "idna (>=3.11,<4.0)",
    "importlib-metadata (>=8.7.0,<9.0.0)",
    "installer (>=0.7.0,<0.8.0)",
    "jaraco-classes (>=3.4.0,<4.0.0)",
    "jaraco-context (>=6.0.1,<7.0.0)",
    "jaraco-functools (>=4.3.0,<5.0.0)",
    "keyring (>=25.6.0,<26.0.0)",
    "more-itertools (>=10.8.0,<11.0.0)",
    "msgpack (>=1.1.2,<2.0.0)",
    "cleo (>=2.1.0,<3.0.0)",
    "crashtest (>=0.4.1,<0.5.0)",
    "packaging (>=25.0,<26.0)",
    "pbs-installer (>=2025.10.28,<2026.0.0)",
    "pkginfo (>=1.12.1.2,<2.0.0.0)",
    "platformdirs (>=4.5.0,<5.0.0)",
    "poetry (>=2.2.1,<3.0.0)",
    "poetry-core (>=2.2.1,<3.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "pyproject-hooks (>=1.2.0,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "python-engineio (>=4.12.3,<5.0.0)",
    "python-socketio (>=5.14.2,<6.0.0)",
    "pywin32-ctypes (>=0.2.3,<0.3.0)",
    "rapidfuzz (>=3.14.1,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "requests-toolbelt (>=1.0.0,<2.0.0)",
    "shellingham (>=1.5.4,<2.0.0)",
    "simple-websocket (>=1.1.0,<2.0.0)",
    "sniffio (>=1.3.1,<2.0.0)",
    "tomli (>=2.3.0,<3.0.0)",
    "tomlkit (>=0.13.3,<0.14.0)",
    "trove-classifiers (>=2025.9.11.17,<2026.0.0.0)",
    "typing-extensions (>=4.15.0,<5.0.0)",
    "urllib3 (>=2.5.0,<3.0.0)",
    "virtualenv (>=20.35.3,<21.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "websocket-client (>=1.9.0,<2.0.0)",
    "wsproto (>=1.2.0,<2.0.0)",
    "zipp (>=3.23.0,<4.0.0)",
    "zstandard (>=0.25.0,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
